+++ inst/doc/Introduction.R
@@ -7,71 +7,54 @@ knitr::opts_chunk$set(
   fig.height = 5
 )
 
-## ---- eval=FALSE--------------------------------------------------------------
-#  install.packages("incidence2")
-
-## ---- eval=FALSE--------------------------------------------------------------
-#  if (!require(remotes)) {
-#    install.packages("remotes")
-#  }
-#  remotes::install_github("reconverse/incidence2", build_vignettes = TRUE)
-
 ## ----data---------------------------------------------------------------------
 library(outbreaks)
 library(incidence2)
-
 dat <- ebola_sim_clean$linelist
 class(dat)
 str(dat)
 
-## ----incid1, fig.height = 5, dpi = 90-----------------------------------------
-i <- incidence(dat, date_index = date_of_onset)
-i
-summary(i)
-plot(i)
+## ----daily, fig.height = 5, dpi = 90------------------------------------------
+daily <- incidence(dat, date_index = date_of_onset)
+daily
+summary(daily)
+plot(daily)
 
-## ----i_7----------------------------------------------------------------------
-
+## ----sevenday-----------------------------------------------------------------
 # 7 day incidence
-i_7 <- incidence(dat, date_index = date_of_onset, interval = 7)
-i_7
-plot(i_7, color = "white")
+seven <- incidence(dat, date_index = date_of_onset, interval = 7)
+seven
+plot(seven, color = "white")
 
-## ----iweekly------------------------------------------------------------------
-
+## ----weekly-------------------------------------------------------------------
 # year-weekly, starting on Monday (ISO week, default)
-i_weekly <- incidence(dat, date_index = date_of_onset, interval = "week")
-plot(i_weekly, color = "white")
+weekly <- incidence(dat, date_index = date_of_onset, interval = "week")
+plot(weekly, color = "white")
 
-
-## ----ilarger------------------------------------------------------------------
+## ----otherinterval------------------------------------------------------------
 # bi-weekly, based on first day in data
-i_biweekly <- incidence(dat, date_index = date_of_onset, interval = "2 weeks")
-plot(i_biweekly, color = "white")
-# year-monthly
-i_monthly <- incidence(dat, date_index = date_of_onset, interval = "month")
-plot(i_monthly, color = "white")
-# year-quarterly
-i_quarterly <- incidence(dat, date_index = date_of_onset, interval = "quarter")
-plot(i_quarterly, color = "white")
-# year-quarterly
-i_year <- incidence(dat, date_index = date_of_onset, interval = "year")
-plot(i_year, color = "white")
+biweekly <- incidence(dat, date_index = date_of_onset, interval = "2 weeks")
+plot(biweekly, color = "white")
+# monthly
+monthly <- incidence(dat, date_index = date_of_onset, interval = "month")
+plot(monthly, color = "white")
+# quarterly
+quarterly <- incidence(dat, date_index = date_of_onset, interval = "quarter")
+plot(quarterly, color = "white")
+# year
+yearly <- incidence(dat, date_index = date_of_onset, interval = "year")
+plot(yearly, color = "white", n.breaks = 2)
 
 ## ----gender-------------------------------------------------------------------
-i_weekly_sex <- incidence(dat, date_index = date_of_onset,
-                     interval = "week", groups = gender)
-i_weekly_sex
-summary(i_weekly_sex)
-
+weekly_grouped <- incidence(dat, date_of_onset, interval = "week", groups = gender)
+weekly_grouped
+summary(weekly_grouped)
 # A singular plot
-plot(i_weekly_sex, fill = gender, color = "white")
-
+plot(weekly_grouped, fill = gender, color = "white")
 # a multi-facet plot
-facet_plot(i_weekly_sex, fill = gender, n_breaks = 5, angle = 45, color = "white")
+facet_plot(weekly_grouped, fill = gender, n.breaks = 5, angle = 45, color = "white")
 
 ## ----outcome_hospital---------------------------------------------------------
-inci <- incidence(dat, date_index = date_of_onset,
-                     interval = "week", groups = c(outcome, hospital, gender))
-facet_plot(inci, facets = hospital, fill = outcome, nrow = 3, n_breaks = 5, angle = 45)
+inci <- incidence(dat, date_of_onset, interval = "week", groups = c(outcome, hospital))
+facet_plot(inci, facets = hospital, fill = outcome, nrow = 3, n.breaks = 5, angle = 45)